Add [vault](https://www.vaultproject.io) image
This is not yet ready for prime time, but it should provide a working
development environment, provided you set the `DEBUG` environment
variable. If DEBUG is not set you can use the ZOOKEEPER_ADDRESS variable
to have a permanent store on zookeeper. Notice that for the moment we do
not have a proper SSL setup so usage in production is not a good idea.