Put our SLC5 container on life support
SLC5 is no longer supported but we still use it for our builds. Now many sites
(including GitHub) use TLSv1.2 which is not supported by the default OpenSSL
version of SLC5: a consequence of this is that all `git clone` commands from
GitHub fail miserably.

This commit adds a custom version of OpenSSL, Curl and Git, picked by default
by every command in the container.