Make packer.json buildable using packer 1.3.4
This is the version packaged on Ubuntu, and thus the version installed by the
GitHub action.

Most changes only make the packer.json files compatible with the build procedure
used in the GitHub action, by specifying file names relative to the packer.json
file.

However, for the two newest Ubuntu images, `"tags": ["latest"]' has to be
replaced with `"tag": "latest"' as Packer 1.3.4 doesn't support the former.