Move active Docker images to CERN registry (#246)
* Allow pushing to CERN registry if packer.json requests it

* Remove obsolete Jenkinsfile

This build script hasn't been used in years, and it doesn't run as-is
on our infrastructure any more. We use GitHub Actions for deploying
individual images instead.

* Move active Docker images to registry.cern.ch

Images that are being used at the moment should be published to
registry.cern.ch instead, as our Docker Hub account's future is
uncertain.